backend/requirements.txt
# === Framework și server ASGI ===
fastapi==0.115.12        # Framework web modern, bazat pe ASGI, pentru API-uri rapide
starlette==0.46.1        # Toolkit ASGI pe care se bazează FastAPI
uvicorn==0.34.0          # Server ASGI performant pentru rularea aplicației

# === Comunicație WebSocket ===
websockets[deflate]==11.0.3  # Client și server WebSocket cu suport pentru compresie

# === Validare și serializare date ===
pydantic==2.11.1          # Validare și serializare a datelor cu modele
pydantic_core==2.33.0     # Motor de bază performant pentru Pydantic
email_validator==2.2.0    # Validare automată de formulă pentru adrese de email

# === Bază de date și migrații ===
SQLAlchemy==2.0.40        # ORM pentru definirea și interogarea modelelor DB
PyMySQL==1.1.1            # Driver MySQL pentru SQLAlchemy
alembic==1.11.0           # Gestionare migrații ale schemei bazei de date

# === Autentificare și securitate ===
passlib==1.7.4            # Hashing de parole (PBKDF2, bcrypt etc.)
bcrypt==4.0.1             # Suport bcrypt pentru hashing securizat
python-jose[cryptography]==3.3.0  # Creare și validare JWT cu backend criptografic

# === AWS SDK și S3 ===
boto3==1.37.23            # SDK Python pentru serviciile AWS
botocore==1.37.23         # Componente de bază pentru boto3
s3transfer==0.11.4        # Gestionare transfer fișiere la/din S3
dnspython==2.7.0          # Rezoluție DNS, necesar de boto3 în anumite cazuri

# === I/O asincron și HTTP ===
anyio==4.9.0              # Abstracție cross-backend pentru I/O asincron
sniffio==1.3.1            # Detectează implementarea I/O async folosită
h11==0.14.0               # Pachet HTTP/1.1 de bază utilizat de Starlette
urllib3==2.3.0            # Client HTTP robust și configurabil
idna==3.10                # Suport puneri în cod pentru domenii internaționale

# === Utilitare generale ===
click==8.1.8              # Construirea de CLI-uri
python-dateutil==2.9.0.post0  # Manipulare flexibilă a datelor și orelor
python-dotenv==1.1.0      # Încărcare variabile de mediu din fișier .env
typing-inspection==0.4.0  # Instrumente pentru introspecția tipurilor la runtime
typing_extensions==4.13.0 # Extensii pentru module de tip Python
six==1.17.0               # Compatibilitate Python 2 și 3
annotated-types==0.7.0    # Extensii de tipuri pentru Pydantic

# === Criptografie low-level ===
cryptography==39.0.1      # Operații criptografice low-level (AES, RSA etc.)

# === Migrare și sincronizare pachete extra ===
websockets[deflate]==11.0.3  # (dublu listat pentru exemplu de grupare)